Added beanconfiguration class and abstract invoice service integration tests
