+++ app/src/test/groovy/pl/futurecollars/invoicing/service/InvoiceServiceInMemoryTest.groovy
@@ -1,0 +1,11 @@
+package pl.futurecollars.invoicing.service
+
+import pl.futurecollars.invoicing.db.Database
+import pl.futurecollars.invoicing.db.memory.InMemoryDatabase
+
+class InvoiceServiceInMemoryTest extends InvoiceServiceAbstractIntegrationTest{
+
+    Database getDatabase() {
+        return new InMemoryDatabase()
+    }
+}
